requirements.txt
django==5.0.3
djangorestframework==3.14.0
drf-yasg==1.21.7